pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradepilot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic-settings>=2.2.1",
  "sqlalchemy>=2.0.29",
  "psycopg[binary]>=3.1.18",
  "httpx>=0.27.0",
  "slack-sdk>=3.27.0",
  "botbuilder-core>=4.15.0",
  "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["tradepilot*"]
